+++ acme/Constants.py
@@ -4,7 +4,7 @@
 #	(c) 2020 by Andreas Kraft
 #	License: BSD 3-Clause License. See the LICENSE file for further details.
 #
-#	Various CSE and oneM2M constante
+#	Various CSE and oneM2M constants
 #
 
 class Constants(object):