+++ acme/resources/Resource.py
@@ -9,6 +9,7 @@
 
 from Logging import Logging
 from Constants import Constants as C
+from Types import BasicType as BT, Cardinality as CAR, RequestOptionality as RO, Announced as AN
 from Configuration import Configuration
 import Utils, CSE
 import datetime, random
@@ -20,17 +21,19 @@ class Resource(object):
 	_srn	= '__srn__'
 	_node	= '__node__'
 
-	def __init__(self, tpe, jsn=None, pi=None, ty=None, create=False, inheritACP=False, readOnly=False, rn=None):
+	def __init__(self, tpe, jsn=None, pi=None, ty=None, create=False, inheritACP=False, readOnly=False, rn=None, attributeDefinitions=None):
 		self.tpe = tpe
 		self.readOnly = readOnly
 		self.inheritACP = inheritACP
 		self.json = {}
+		self.attributeDefinitions = attributeDefinitions
 
 		if jsn is not None: 
 			if tpe in jsn:
 				self.json = jsn[tpe].copy()
 			else:
 				self.json = jsn.copy()
+			self._originalJson = jsn.copy()	# keep for validation later
 		else:
 			pass
 			# TODO Exception?
@@ -103,6 +106,13 @@ class Resource(object):
 	# Implemented in sub-classes.
 	def activate(self, parentResource, originator):
 		Logging.logDebug('Activating resource: %s' % self.ri)
+
+		# validate the attributes
+		if not (result := self.validateAttributes(create=True, jsn=self._originalJson))[0]:
+			return result
+
+
+		# validate the resource logic
 		if not (result := self.validate(originator, create=True))[0]:
 			return result
 
@@ -152,6 +162,11 @@ class Resource(object):
 	# Update this resource with (new) fields.
 	# Call validate() afterward to react on changes.
 	def update(self, jsn=None, originator=None):
+
+		# validate the attributes
+		if not (result := self.validateAttributes(create=False, jsn=jsn))[0]:
+			return result
+
 		if jsn is not None:
 			if self.tpe not in jsn:
 				Logging.logWarn("Update types don't match")
@@ -194,25 +209,24 @@ class Resource(object):
 		CSE.notification.checkSubscriptions(self, C.netCreateDirectChild, childResource)
 
 
-	# Child was removed from the resource.
 	def childRemoved(self, childResource, originator):
+		""" Call when child resource was removed from the resource. """
 		CSE.notification.checkSubscriptions(self, C.netDeleteDirectChild, childResource)
 
 
-	# MUST be implemented by each class
 	def canHaveChild(self, resource):
+		""" MUST be implemented by each class."""
 		raise NotImplementedError('canHaveChild()')
 
 
-	# Is be called from child class
 	def _canHaveChild(self, resource, allowedChildResourceTypes):
+		""" It checks whether a fresource may have a certain child resources. This is called from child class. """
 		from .Unknown import Unknown # Unknown imports this class, therefore import only here
 		return resource['ty'] in allowedChildResourceTypes or isinstance(resource, Unknown)
 
 
-	# Validate a resource. Usually called within activate() or
-	# update() methods.
 	def validate(self, originator=None, create=False):
+		""" Validate a resource. Usually called within activate() or update() methods. """
 		Logging.logDebug('Validating resource: %s' % self.ri)
 		if (not Utils.isValidID(self.ri) or
 			not Utils.isValidID(self.pi) or
@@ -222,11 +236,61 @@ class Resource(object):
 		return (True, C.rcOK)
 
 
-	# Validate possible expirations, of self or child resources.
-	# MAY be implemented by child class
 	def validateExpirations(self):
+		"""	Validate possible expirations, of self or child resources.
+			MAY be implemented by child class.
+		"""
 		pass
 
+
+	def	validateAttributes(self, create=True, jsn=None):
+		Logging.logDebug('Validating attributes for resource: %s' % self.ri)
+
+		""" Validate a resources attributes for types etc."""
+		if self.attributeDefinitions is None:
+			return (True, C.rcOK)
+
+		# Either take the provided jsn or self.jsn
+		jsn = self.json if jsn is None else jsn
+
+		# determine the request column, depending on create or updates
+		reqp = 2 if create else 3
+
+		for r, p in self.attributeDefinitions.items():
+
+			# Check whether the attribute is allowed or mandatory in the request
+			if (v := jsn.get(r)) is None:
+				if p[reqp] == RO.M:		# Not okay, this attribute is mandatory
+					Logging.logDebug('Cannot find mandatory attribute: %s' % r)
+					return (False, C.rcBadRequest)
+				if p[reqp] in [ RO.NP, RO.O]:	# Okay that the attribute is not in the json, since it is provided or optional
+					continue
+			else:
+				if p[reqp] == RO.NP:
+					Logging.logDebug('Found non-provision attribute: %s' % r)
+					return (False, C.rcBadRequest)
+
+			# Check whether the value is of the correct type
+			pt = p[0]	# type
+			pc = p[1]	# cardinality
+			if pt == BT.positiveInteger and isinstance(v, int) and v > 0:
+				continue
+			if pt == BT.nonNegInteger and isinstance(v, int) and v >= 0:
+				continue
+			if pt in [ BT.string, BT.timestamp, BT.anyURI ] and isinstance(v, str):
+				continue
+			if pt == BT.list and isinstance(v, list):
+				continue
+			if pt == BT.boolean and isinstance(v, bool) and v > 0:
+				continue
+			if pt == BT.geoCoordinates and isinstance(v, dict):
+				continue
+			
+			# fall-through means: not validated
+			Logging.logDebug('Attribute/value validation error: %s=%s' % (r, str(v)))
+			return (False, C.rcBadRequest)
+
+		return (True, C.rcOK)